Move edit hearing logic to NewHearingJudgeEvent
