+++ service/src/main/java/uk/gov/hmcts/reform/fpl/events/judicial/NewHearingJudgeEvent.java
@@ -6,6 +6,8 @@ import lombok.RequiredArgsConstructor;
 import uk.gov.hmcts.reform.fpl.model.CaseData;
 import uk.gov.hmcts.reform.fpl.model.HearingBooking;
 
+import java.util.Optional;
+
 @Data
 @Builder(toBuilder = true)
 @RequiredArgsConstructor
@@ -13,5 +15,6 @@ public class NewHearingJudgeEvent {
 
     private final HearingBooking hearing;
     private final CaseData caseData;
+    private final Optional<HearingBooking> oldHearing;
 
 }
