+++ service/src/main/java/uk/gov/hmcts/reform/fpl/handlers/judicial/HandleHearingModificationRolesEventHandler.java
@@ -40,28 +40,4 @@ public class HandleHearingModificationRolesEventHandler {
             log.error("Error when handling roles on cancelled hearings", e);
         }
     }
-
-    @EventListener
-    public void handleEditedHearingRoles(final HandleHearingModificationRolesEvent event) {
-        // Not an async function as it has to take place before we grant more roles, in case the times overlap
-        // when relisting. it has to be caught though to make sure nothing else afterward is impacted in case of
-        // failure
-
-        try {
-            event.getCaseData().getAllHearings()
-                .forEach(hearing -> {
-                    Optional<Element<HearingBooking>> oldHearing = findElement(hearing.getId(),
-                        event.getCaseDataBefore().getAllHearings());
-                    if (oldHearing.isPresent() && !oldHearing.get().equals(hearing)) {
-                        // hearing that was present beforehand, and has changed in some way
-                        judicialService.deleteSpecificHearingRole(event.getCaseData().getId(),
-                            oldHearing.get().getValue());
-                    }
-                });
-        } catch (Exception e) {
-            log.error("Error when handling roles on cancelled hearings", e);
-        }
-    }
-
-
 }
