+++ service/src/test/java/uk/gov/hmcts/reform/fpl/handlers/judicial/HandleHearingModificationRolesEventHandlerTest.java
@@ -54,64 +54,4 @@ class HandleHearingModificationRolesEventHandlerTest {
         verify(judicialService).deleteSpecificHearingRole(12345L, newCancelledBooking);
         verifyNoMoreInteractions(judicialService);
     }
-
-    @Test
-    void shouldHandleEditedHearingRolesWhenTimeChanges() {
-        LocalDateTime now = LocalDateTime.now();
-        UUID hearingId = UUID.randomUUID();
-
-        HearingBooking oldHearing = HearingBooking.builder()
-            .startDate(now)
-            .judgeAndLegalAdvisor(JudgeAndLegalAdvisor.builder()
-                .judgeJudicialUser(JudicialUser.builder()
-                    .idamId("idamId")
-                    .build())
-                .build())
-            .build();
-
-        HearingBooking newHearing = oldHearing.toBuilder()
-            .startDate(now.plusDays(2))
-            .build();
-
-        underTest.handleEditedHearingRoles(new HandleHearingModificationRolesEvent(
-            CaseData.builder().id(12345L).hearingDetails(List.of(element(hearingId, newHearing))).build(),
-            CaseData.builder().id(12345L).hearingDetails(List.of(element(hearingId, oldHearing))).build()
-        ));
-
-        verify(judicialService).deleteSpecificHearingRole(12345L, oldHearing);
-        verifyNoMoreInteractions(judicialService);
-    }
-
-    @Test
-    void shouldHandleEditedHearingRolesWhenJudgeChanges() {
-        LocalDateTime now = LocalDateTime.now();
-        UUID hearingId = UUID.randomUUID();
-
-        HearingBooking oldHearing = HearingBooking.builder()
-            .startDate(now)
-            .judgeAndLegalAdvisor(JudgeAndLegalAdvisor.builder()
-                .judgeJudicialUser(JudicialUser.builder()
-                    .idamId("idamId")
-                    .build())
-                .build())
-            .build();
-
-        HearingBooking newHearing = oldHearing.toBuilder()
-            .judgeAndLegalAdvisor(JudgeAndLegalAdvisor.builder()
-                .judgeJudicialUser(JudicialUser.builder()
-                    .idamId("idamId2")
-                    .build())
-                .build())
-            .build();
-
-        underTest.handleEditedHearingRoles(new HandleHearingModificationRolesEvent(
-            CaseData.builder().id(12345L).hearingDetails(List.of(element(hearingId, newHearing))).build(),
-            CaseData.builder().id(12345L).hearingDetails(List.of(element(hearingId, oldHearing))).build()
-        ));
-
-        verify(judicialService).deleteSpecificHearingRole(12345L, oldHearing);
-        verifyNoMoreInteractions(judicialService);
-    }
-
-
 }
